Change: Making my code more OO
I add a Parser into my code so that in the
NUSYapBot.java, I don't have to add to the long
if-else chain every time a new command is added.
I extract the process of checking the command
into Parser, and create Command classes for
each of the different commands. Currently halfway
done, Command classes to be created:
-markTaskCommand
-unmarkTaskCommand
-deleteCommand
-findCommand

Other changes:
- change display format of Task: now the date
is shown in more readable way.
- delete Ui function of printGoodbyeMessage:
the function is served by ByeCommand